Playtesting feedback round 3 changes
1. Playtester handled game challenges well, and was not tripped-up by previous points of confuion (thanks to the round 1 and 2 changes made)
They successfully completed the game with only needing a hint for level 4. Noted this was significantly more difficult than level 5, and was the hardest by far.
This level was also the biggest challenge by a notable amount for round 1 and 2

This feedback has been taken into consideration by switching Scenes "Level4" and "Level5" around
They have exchanged names and order in the Scene List.

This gives the game a more natural order of progression.

2. No other notable points were raised during this feedback session.
Game should be ready for prototype final build, if no issues are raised by other internal team members